main refact
